+++ tools/tensorium-miner/stratum_client.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "common.h"
+#ifdef __cplusplus
+extern "C" {
+#endif
+void stratum_client_run(const MinerConfig *cfg, SharedState *state);
+#ifdef __cplusplus
+}
+#endif